First commit as actual Python package
