+++ .gitignore
@@ -1,3 +1,7 @@
+*egg-info
+*.egg
 __pycache__
-test/
+build
+dist
+MANIFEST
 .idea
