+++ MANIFEST.in
@@ -1,0 +1,2 @@
+graft sweet_potato/data
+graft sweet_potato/examples
