+++ sweetpotato/__init__.py
@@ -1,0 +1,2 @@
+from .cli import parse_args as main
+from .common import AUTHOR_EMAIL, AUTHOR_NAME, DESCRIPTION, LICENSE, PROGNAME, VERSION
