+++ sweetpotato/core.py
@@ -1,14 +1,6 @@
-#!/usr/bin/env python3
-"""A tool for managing Minecraft servers on GNU/Linux."""
-import argparse
-try:
-    import bottle
-except ImportError:
-    bottle = None
 import configparser
 import json
 # import logging
-import os
 import pty
 import shlex
 import subprocess
@@ -18,117 +10,27 @@ import tarfile
 import urllib.error
 import urllib.request
 
-from collections import OrderedDict
 from datetime import datetime
-try:
-    from markdown import markdown
-except ImportError:
-    markdown = None
-from threading import Thread
+from .common import *
+from .daemon import Daemon
+from .error import *
 
 
-__author__ = 'Hristos N. Triantafillou <me@hristos.triantafillou.us>'
-__license__ = 'GPLv3'
-__mcversion__ = '1.8.1'
-__progname__ = 'sweetpotato'
-__version__ = '0.34.15b'
+__all__ = ['SweetpotatoConfig', 'create_server', 'dependency_check', 'die_silently', 'error_and_die',
+           'get_exe_path', 'get_jar', 'get_uptime', 'get_uptime_raw', 'get_uptime_string', 'is_server_running',
+           'list_players', 'list_players_as_list', 'read_conf_file', 'reread_settings', 'restart_server',
+           'run_server_backup', 'send_command', 'start_server', 'stop_server', 'validate_directories',
+           'validate_mem_values', 'validate_settings']
 
 
-BASE_DIR = os.path.dirname(os.path.realpath(__file__))
-DEFAULT_PERMGEN = '256'
-DEFAULT_SCREEN_NAME = 'SweetpotatoWorld'
-DEFAULT_SERVER_PORT = '25565'
-DEFAULT_WEBUI_PORT = 8080
-DEFAULT_WORLD_NAME = DEFAULT_SCREEN_NAME
-DESCRIPTION = "Manage your Minecraft server on a GNU/Linux system."
-HOME_DIR = os.getenv('HOME')
-CONFIG_DIR = '{0}/.config/{1}'.format(HOME_DIR, __progname__)
-DEFAULT_LOG_DIR = os.path.join(CONFIG_DIR, 'logs')
-DEFAULT_CONF_FILE = '{0}/{1}.conf'.format(CONFIG_DIR, __progname__)
-REQUIRED = 'backup_dir mem_format mem_max mem_min port screen_name server_dir world_name'
-SERVER_WAIT_TIME = 0.5
-FORGE_DL_URL = 'http://files.minecraftforge.net/maven/net/minecraftforge/forge/{0}/{1}'
-FORGE_JAR_NAME = 'forge-{}-universal.jar'
-VANILLA_DL_URL = 'https://s3.amazonaws.com/Minecraft.Download/versions/{0}/{1}'
-VANILLA_JAR_NAME = 'minecraft_server.{}.jar'
-
-
-class Colors:
-    blue = '\033[34m'
-    light_blue = '\033[94m'
-    yellow_green = '\033[92m'
-    green = '\033[32m'
-    light_red = '\033[91m'
-    red = '\033[31m'
-    yellow = '\033[33m'
-    end = '\033[0m'
-
-
-class SweetpotatoIOErrorBase(IOError):
-    def __init__(self, msg):
-        self.msg = msg
-
-    def __str__(self):
-        return repr(self.msg)
-
-
-class BackupFileAlreadyExistsError(SweetpotatoIOErrorBase):
-    """Raised when the target backup file already exists."""
-    pass
-
-
-class ConfFileError(SweetpotatoIOErrorBase):
-    """Raised when a given conf file doesn't exist or have the right section."""
-    pass
-
-
-class EmptySettingError(SweetpotatoIOErrorBase):
-    """Raised when a required setting value is None."""
-    pass
-
-
-class MinimalInstallError(SweetpotatoIOErrorBase):
-    """Raised when we are running as a minimal install."""
-    pass
-
-
-class MissingExeError(SweetpotatoIOErrorBase):
-    """Raised when a required executable is missing."""
-    pass
-
-
-class NoDirFoundError(SweetpotatoIOErrorBase):
-    """Raised when a given directory does not exist."""
-    pass
-
-
-class NoJarFoundError(SweetpotatoIOErrorBase):
-    """Raised when a given jar file does not exist."""
-    pass
-
-
-class ServerAlreadyRunningError(SweetpotatoIOErrorBase):
-    """Raised when the configured server is already running."""
-    pass
-
-
-class ServerNotRunningError(SweetpotatoIOErrorBase):
-    """Raised when the server is not running but was expected to be."""
-    pass
-
-
-class UnsupportedVersionError(SweetpotatoIOErrorBase):
-    """Raised when we try to generate a server.properties for an unsupported mc_version."""
-    pass
-
-
-class SweetpotatoConfig:
+class SweetpotatoConfig(Daemon):
     """
     Representing the settings sweetpotato needs to function.
     """
-    header = "## Generated by {0} at {1}".format(__progname__, datetime.now().strftime('%c'))
+    header = "## Generated by {0} at {1}".format(PROGNAME, datetime.now().strftime('%c'))
 
     def __init__(self):
+        super().__init__(DEFAULT_PIDFILE)
         self.backup_dir = None
         self.compression = 'gz'
         self.conf_file = None
@@ -139,7 +41,7 @@ class SweetpotatoConfig:
         self.mem_format = None
         self.mem_max = None
         self.mem_min = None
-        self.mc_version = __mcversion__
+        self.mc_version = MCVERSION
         self.permgen = None
         self.port = DEFAULT_SERVER_PORT
         self.running = False
@@ -163,6 +65,7 @@ class SweetpotatoConfig:
 
     @property
     def as_json(self):
+        self.__dict__.pop('pidfile')  # TODO: maybe not do this?
         try:
             players = list_players_as_list(list_players(self))
             raw = get_uptime_raw(self.server_dir, self.world_name, False)
@@ -278,7 +181,7 @@ motd=Welcome to {0}!
             )
 
 
-def agree_to_eula(eula_txt, force, print_pre, quiet):
+def _agree_to_eula(eula_txt, force, print_pre, quiet):
     """
     Checks for a eula.txt file in server, and verifies it contains 'eula=true'.
 
@@ -306,7 +209,7 @@ def agree_to_eula(eula_txt, force, print_pre, quiet):
     return True
 
 
-def can_xz():
+def _can_xz():
     if sys.version_info[1] < 3:
         return False
     else:
@@ -325,7 +228,7 @@ def create_server(settings, quiet):
     """
     print_pre = '[' + Colors.yellow_green + 'create' + Colors.end + '] '
     backup_dir = settings.backup_dir
-    force = is_forced(settings)
+    force = _is_forced(settings)
     server_dir = settings.server_dir
     world_name = settings.world_name
 
@@ -354,19 +257,19 @@ def create_server(settings, quiet):
         if not quiet:
             print(print_pre + Colors.light_blue + 'Creating {} ...'.format(backup_dir), end=' ')
             sys.stdout.flush()
-            list_or_create_dir(backup_dir)
+            _create_dir(backup_dir)
             print('Done!' + Colors.end)
         else:
-            list_or_create_dir(backup_dir)
+            _create_dir(backup_dir)
 
     if not os.path.isdir(server_dir):
         if not quiet:
             print(print_pre + Colors.light_blue + 'Creating {} ...'.format(server_dir), end=' ')
             sys.stdout.flush()
-            list_or_create_dir(server_dir)
+            _create_dir(server_dir)
             print('Done!' + Colors.end)
         else:
-            list_or_create_dir(server_dir)
+            _create_dir(server_dir)
     elif not quiet:
             print(print_pre + Colors.light_blue + 'Found {}!'.format(server_dir) + Colors.end)
 
@@ -393,7 +296,7 @@ def create_server(settings, quiet):
             die_silently()
 
     eula_txt = os.path.join(server_dir, 'eula.txt')
-    agree_to_eula(eula_txt, force, print_pre, quiet)
+    _agree_to_eula(eula_txt, force, print_pre, quiet)
 
     server_properties = os.path.join(server_dir, 'server.properties')
     write_server_properties(print_pre, server_properties, settings, quiet)
@@ -429,7 +332,7 @@ def error_and_die(msg):
     sys.exit(1)
 
 
-def format_launch_cmd(settings):
+def _format_launch_cmd(settings):
     """
     Reads from settings to return the proper command string for launching
     the configured server.
@@ -461,7 +364,7 @@ def get_exe_path(exe):
     return byte_output.decode().strip() or None
 
 
-def get_java_procs():
+def _get_java_procs():
     """Checks for running Java processes and return the cwd, exe, and pid of any we find."""
     cwd, exe, pid = None, None, None
     p = subprocess.Popen(['pgrep', 'java'], stdout=subprocess.PIPE)
@@ -609,7 +512,7 @@ def get_uptime_string(uptime):
     )
 
 
-def is_forced(settings):
+def _is_forced(settings):
     try:
         force = settings.force
     except AttributeError:
@@ -617,7 +520,7 @@ def is_forced(settings):
     return force
 
 
-def is_screen_started(screen_name):
+def _is_screen_started(screen_name):
     """
     Checks the given screen name for a running session,
     returning the PID and name in a 'PID.NAME' formatted string.
@@ -658,7 +561,7 @@ def is_server_running(server_dir_name):
     @param server_dir_name:
     @return:
     """
-    java_procs = get_java_procs()
+    java_procs = _get_java_procs()
     server_dir = server_dir_name.rstrip('/')
     if isinstance(java_procs, tuple):
         # just one java proc
@@ -684,7 +587,7 @@ def is_server_running(server_dir_name):
     return False
 
 
-def list_or_create_dir(d):
+def _create_dir(d):
     """
     Returns a list of config files in the config directory, creating the folder if it does not exist.
 
@@ -695,7 +598,6 @@ def list_or_create_dir(d):
             os.makedirs(d)
         except PermissionError as e:
             error_and_die(e)
-    return os.listdir(d)
 
 
 def list_players(settings):
@@ -844,7 +746,7 @@ def start_server(print_pre, settings, quiet):
     server_dir = settings.server_dir
     world_name = settings.world_name
 
-    screen_started = is_screen_started(screen_name)
+    screen_started = _is_screen_started(screen_name)
     server_running = is_server_running(server_dir)
 
     if settings.forge:
@@ -931,10 +833,10 @@ def restart_server(print_pre, settings, quiet):
     server_dir = settings.server_dir
     world_name = settings.world_name
 
-    screen_started = is_screen_started(screen_name)
+    screen_started = _is_screen_started(screen_name)
     server_running = is_server_running(server_dir)
 
-    launch_server = format_launch_cmd(settings)
+    launch_server = _format_launch_cmd(settings)
 
     if not screen_started:
         start_screen(screen_name, server_dir)
@@ -964,10 +866,10 @@ def run_server_backup(print_pre, settings, quiet, offline=False):
     """
     backup_dir = settings.backup_dir
     date_stamp = datetime.now().strftime('%Y-%m-%d')
-    force = is_forced(settings)
+    force = _is_forced(settings)
     forge = settings.forge
 
-    launch_server = format_launch_cmd(settings)
+    launch_server = _format_launch_cmd(settings)
 
     screen_name = settings.screen_name
     server_dir = settings.server_dir
@@ -1011,9 +913,9 @@ def run_server_backup(print_pre, settings, quiet, offline=False):
         sys.stdout.flush()
         send_command('say Server backing up now', screen_name)
 
-    list_or_create_dir(backup_dir)
+    _create_dir(backup_dir)
 
-    if not can_xz():
+    if not _can_xz():
         compression = 'gz'
 
     tar = tarfile.open(full_path_to_backup_file, 'w:{}'.format(compression))
@@ -1027,7 +929,7 @@ def run_server_backup(print_pre, settings, quiet, offline=False):
         send_command('say Backup complete', screen_name)
         send_command('save-on', screen_name)
 
-    screen_started = is_screen_started(screen_name)
+    screen_started = _is_screen_started(screen_name)
     if offline and server_running_at_first:
         if not screen_started:
             start_screen(screen_name, server_dir)
@@ -1037,239 +939,6 @@ def run_server_backup(print_pre, settings, quiet, offline=False):
         send_command(launch_server, screen_name)
     if not quiet:
         print('Done!' + Colors.end)
-
-
-def run_webui(settings, quiet):
-    """
-    Run the WebUI or print a message about why not.
-
-    @param settings:
-    @return:
-    """
-    # TODO: decorator for DRY view context
-    # TODO: POST routes for control functions
-    if bottle:
-        app = bottle.app()
-        static_path = os.path.join(BASE_DIR, 'data/static')
-        tpl_path = os.path.join(BASE_DIR, 'data/tpl')
-
-        if not os.path.isdir(static_path) or not os.path.isdir(tpl_path):
-            raise MinimalInstallError(
-                "The WebUI cannot run while using a minimal install. Exiting ..."
-            )
-
-        bottle.debug(False)
-        bottle.TEMPLATE_PATH.insert(0, tpl_path)
-
-        if markdown:
-            def readme_md_to_html(md_file):
-                """
-                Runs markdown() on a markdown-formatted README.md file to generate html.
-
-                @param md_file:
-                @return:
-                """
-                html = "% rebase('base.tpl', title='README.md')\n"
-                m = open(md_file, 'r')
-                for l in m.readlines():
-                    html += markdown(l)
-                m.close()
-                return html
-
-            @bottle.route('/readme')
-            def readme():
-                path = bottle.request.path
-                html = readme_md_to_html(os.path.join(BASE_DIR, 'README.md'))
-                return bottle.template(
-                    html,
-                    path=path,
-                    __version__=__version__)
-        else:
-            @bottle.route('/readme')
-            @bottle.view('readme_no_md')
-            def readme():
-                path = bottle.request.path
-                return {'path': path, '__version__': __version__}
-
-        @bottle.get('/backup')
-        @bottle.post('/backup')
-        @bottle.view('backup')
-        def backups():
-            s = reread_settings(settings)
-            offline = None
-            is_running = is_server_running(s.server_dir)
-            path = bottle.request.path
-            request_method = bottle.request.method
-            todays_file = '{0}_{1}.tar.{2}'.format(
-                datetime.now().strftime('%Y-%m-%d'),
-                s.world_name,
-                s.compression
-            )
-
-            backup_dir_contents = os.listdir(s.backup_dir)
-            unsorted_backup_file_list = []
-            for backup_file in backup_dir_contents:
-                full_path_to_backup_file = os.path.join(s.backup_dir, backup_file)
-                raw_backup_file_size = os.path.getsize(full_path_to_backup_file)
-                if os.path.isfile(full_path_to_backup_file):
-                    backup_file_size = raw_backup_file_size / 1000000
-                    dc = len(str(backup_file_size).split('.')[-1])
-                    dc_diff = dc - 2
-                    try:
-                        trimmed_size = float(str(backup_file_size)[0:-dc_diff])
-                    except ValueError:
-                        trimmed_size = backup_file_size
-                    unsorted_backup_file_list.append({
-                        'bit': 'MB',
-                        'file': backup_file,
-                        'size': trimmed_size
-                    })
-            backup_file_list = sorted(unsorted_backup_file_list, key=lambda k: k['size'], reverse=True)
-
-            if bottle.request.method == 'POST':
-                postdata = bottle.request.POST
-                offline = 'offline' in postdata
-                t = Thread(
-                    target=run_server_backup,
-                    args=('', s, True),
-                    kwargs={'offline': offline})
-                t.daemon = True
-                t.start()
-
-            return {
-                'backup_dir_contents': backup_dir_contents,
-                'backup_file_list': backup_file_list,
-                'offline': offline,
-                'path': path,
-                'request_method': request_method,
-                'server_running': is_running,
-                'todays_file': todays_file,
-                'world_name': s.world_name,
-                '__version__': __version__
-            }
-
-        @bottle.route('/')
-        @bottle.view('index')
-        def index():
-            s = reread_settings(settings)
-            is_running = is_server_running(s.server_dir)
-            path = bottle.request.path
-            pid = None
-            players = None
-            try:
-                raw = get_uptime_raw(s.server_dir, s.world_name, False)
-                u = get_uptime(raw)
-                uptime = get_uptime_string(u)
-                if is_running:
-                    pid = is_running.get('pid')
-                    players = list_players_as_list(list_players(s))
-            except ServerNotRunningError:
-                uptime = None
-            sorted_settings = OrderedDict(sorted(s.__dict__.items()))
-
-            return {
-                'path': path,
-                'pid': pid,
-                'players': players,
-                'server_running': is_running,
-                'sorted_settings': sorted_settings,
-                'uptime': uptime,
-                'world_name': s.world_name,
-                '__version__': __version__
-            }
-
-        @bottle.route('/json')
-        def as_json():
-            bottle.response.content_type = 'application/json'
-            s = reread_settings(settings)
-            s.running = is_server_running(s.server_dir)
-            return s.as_json
-
-        @bottle.get('/server')
-        @bottle.post('/server')
-        @bottle.view('server')
-        def server():
-            s = reread_settings(settings)
-            is_running = is_server_running(s.server_dir)
-            path = bottle.request.path
-            request_method = bottle.request.method
-            restart = None
-            start = None
-            stop = None
-            if bottle.request.method == 'POST':
-                postdata = bottle.request.POST
-                restart = postdata.get('restart')
-                start = postdata.get('start')
-                stop = postdata.get('stop')
-                if restart is not None:
-                    t = Thread(target=restart_server, args=('', s, True))
-                    t.daemon = True
-                    t.start()
-                elif start is not None:
-                    t = Thread(target=start_server, args=('', s, True))
-                    t.daemon = True
-                    t.start()
-                elif stop is not None:
-                    t = Thread(target=stop_server,
-                               args=('', s.screen_name, s.server_dir, s.world_name, True))
-                    t.daemon = True
-                    t.start()
-            return {
-                'path': path,
-                'request_method': request_method,
-                'restart': restart,
-                'start': start,
-                'stop': stop,
-                'server_running': is_running,
-                'world_name': s.world_name,
-                '__version__': __version__
-            }
-
-        # noinspection PyUnresolvedReferences
-        @bottle.route('/backups/<file_path:path>')
-        def serve_backup(file_path):
-            s = reread_settings(settings)
-            return bottle.static_file(file_path, root=s.backup_dir)
-
-        # noinspection PyUnresolvedReferences
-        @bottle.route('/static/<file_path:path>')
-        def serve_static(file_path):
-            return bottle.static_file(file_path, root=static_path)
-
-        @bottle.error(404)
-        @bottle.view('404')
-        def error404(error):
-            path = bottle.request.path
-            return {
-                'error': error,
-                'path': path,
-                '__version__': __version__
-            }
-
-        @bottle.error(500)
-        @bottle.view('500')
-        def error500(error):
-            path = bottle.request.path
-            return {
-                'error': error,
-                'path': path,
-                '__version__': __version__
-            }
-        try:
-            if not quiet:
-                print(Colors.green + '{0} {1} - launching WebUI now!'.format(__progname__, __version__))
-                sys.stdout.flush()
-            bottle.run(app=app, port=settings.webui_port, quiet=quiet)
-        except OSError:
-            if not quiet:
-                error_and_die('Port {} is already in use! Exiting ...\n'.format(settings.webui_port))
-            else:
-                die_silently()
-    elif not quiet:
-        error_and_die('The web component requires both bottle.py to function, '
-                      'with Python-Markdown as an optional dependency.')
-    else:
-        die_silently()
 
 
 def validate_directories(*dirs):
@@ -1380,270 +1049,3 @@ def write_server_properties(print_pre, file, settings, quiet):
         f.close()
     else:
         do_the_write()
-
-
-# def write_to_logfile(msg, settings):
-#     """
-#     Write the supplied message to the configured log.
-#
-#     @param msg:
-#     @param settings:
-#     @return:
-#     """
-#     pass
-
-
-def arg_parse(argz):
-    parser = argparse.ArgumentParser(description=DESCRIPTION, prog=__progname__)
-
-    actions = parser.add_mutually_exclusive_group(required=True)
-    actions.add_argument('-b', '--backup', action='store_true', help='back up your Minecraft server (live)')
-    actions.add_argument('-C', '--create', action='store_true', help='create a server from settings')
-    actions.add_argument('-g', '--genconf', action='store_true', help='generate conf file from passed-in CLI arguments')
-    actions.add_argument('-j', '--json', action='store_true', help='output settings as json')
-    actions.add_argument('-l', '--list', action='store_true', help='list logged-in players')
-    actions.add_argument('-o', '--offline', action='store_true', help='make an offline backup (stops the server)')
-    actions.add_argument('-r', '--restart', action='store_true', help='restart the server')
-    actions.add_argument('--say', help=argparse.SUPPRESS)
-    actions.add_argument('--start', action='store_true', help='start the server in a screen session')
-    # actions.add_argument('--save', action='store_true', help='save the current config')
-    actions.add_argument('--stop', action='store_true', help='stop the server')
-    # actions.add_argument('--testing', action='store_true', help=argparse.SUPPRESS)
-    actions.add_argument('-U', '--uptime', action='store_true', help='Show server uptime in hours')
-    actions.add_argument('-W', '--web', action='store_true', help='run the WebUI')
-    # actions.add_argument('--wipe', action='store_true', help='Wipe configs')
-
-    settings = parser.add_argument_group('Settings', 'config options for %(prog)s')
-    settings.add_argument('-c', '--conf', help='config file containing your settings', metavar='CONF FILE')
-    settings.add_argument('-d', '--backup-dir', help='the FULL path to your backups folder',
-                          metavar='/path/to/backups')
-    settings.add_argument('-x', '--fancy', action='store_true', help="print json with fancy indentation and sorting")
-    settings.add_argument('-F', '--force', help='forces writing of server files, even when they already exist',
-                          action='store_true')
-    settings.add_argument('--level-seed', '--seed', metavar="LEVEL SEED",
-                          help='optional and only applied during world creation')
-    # settings.add_argument('--log-dir', '-L', metavar="LOG DIR",
-    #                       help='set the log location. Default: {}'.format(DEFAULT_LOG_DIR))
-    settings.add_argument('-p', '--port',
-                          help='port you wish to run your server on. Default: {}'.format(DEFAULT_SERVER_PORT))
-    settings.add_argument('-q', '--quiet', action='store_true', help='Silence all output')
-    settings.add_argument('-s', '--server-dir', metavar='/path/to/server',
-                          help='set the FULL path to the directory containing your server files')
-    settings.add_argument('-S', '--screen', metavar='SCREEN NAME',
-                          help='set the name of your screen session. Default: the same as your world')
-    settings.add_argument('-v', '--mc-version', metavar='MC VERSION',
-                          help='set the version of minecraft. Default: The latest stable')
-    settings.add_argument('--webui-port', dest='webui_port',
-                          help='Port to bind to for the WebUI. Default: {}'.format(DEFAULT_WEBUI_PORT))
-    settings.add_argument('-w', '--world', metavar='WORLD NAME',
-                          help='set the name of your Minecraft world. Default: {}'.format(DEFAULT_WORLD_NAME))
-    settings.add_argument('-z', '--compression', choices=['bz2', 'gz', 'xz'], dest='compression',
-                          help='select compression type. Default: gz')
-
-    mem_values = settings.add_mutually_exclusive_group()
-    mem_values.add_argument('-gb', '-GB', help='set min/max memory usage (in gigabytes)',
-                            metavar=('MIN', 'MAX'), nargs=2)
-    mem_values.add_argument('-mb', '-MB', help='set min/max memory usage (in megabytes)',
-                            metavar=('MIN', 'MAX'), nargs=2)
-
-    forge_settings = parser.add_argument_group('Mods', 'Options for Forge users')
-    forge_settings.add_argument('-f', '--forge', help='version of Forge you are using.', metavar='FORGE VERSION')
-    forge_settings.add_argument('-P', '--permgen',
-                                help='Amount of permgen to use in MB. Default: {}'.format(DEFAULT_PERMGEN))
-
-    parser.add_argument('-V', '--version', action='version', version='%(prog)s {0}'.format(__version__))
-
-    args = parser.parse_args(argz)
-    s = SweetpotatoConfig()
-
-    # Read a passed-in conf file
-    if args.conf:
-        try:
-            read_conf_file(args.conf, s)
-        except ConfFileError as e:
-            error_and_die(e)
-        s.conf_file = os.path.realpath(args.conf)
-    # Or try to read the default conf file
-    elif not args.conf:
-        try:
-            read_conf_file(DEFAULT_CONF_FILE, s)
-            s.conf_file = DEFAULT_CONF_FILE
-        except (configparser.NoSectionError, ConfFileError):
-            pass
-    if args.backup_dir:
-        s.backup_dir = args.backup_dir
-    if args.fancy:
-        s.fancy = True
-    if args.force:
-        s.force = args.force
-    if args.forge:
-        s.forge = args.forge
-    if args.mb:
-        s.mem_format = 'MB'
-        s.mem_max = args.mb[1]
-        s.mem_min = args.mb[0]
-    elif args.gb:
-        s.mem_format = 'GB'
-        s.mem_max = args.gb[1]
-        s.mem_min = args.gb[0]
-    if args.level_seed:
-        s.level_seed = args.level_seed
-    if args.permgen:
-        s.permgen = args.permgen
-    if args.port:
-        s.port = args.port
-    if args.quiet:
-        quiet = True
-    else:
-        quiet = False
-    if args.server_dir:
-        s.server_dir = args.server_dir
-    if args.screen:
-        s.screen_name = args.screen
-    if args.mc_version:
-        s.mc_version = args.mc_version
-    if args.webui_port:
-        s.webui_port = args.webui_port
-    if args.world:
-        s.world_name = args.world
-    if args.compression:
-        s.compression = args.compression
-
-    if s.forge:
-        s.mc_version = s.forge.split('-')[0]
-        if not s.permgen:
-            s.permgen = DEFAULT_PERMGEN
-
-    try:
-        validate_settings(s)
-        if s.world_name != DEFAULT_WORLD_NAME and not args.screen and s.screen_name == DEFAULT_WORLD_NAME:
-            s.screen_name = s.world_name
-    except EmptySettingError as e:
-        error_and_die(e)
-
-    try:
-        validate_directories(s.backup_dir, s.server_dir)
-    except NoDirFoundError as e:
-        if not args.create and not args.genconf:
-            error_and_die(e)
-
-    try:
-        validate_mem_values(s.mem_min, s.mem_max)
-    except ValueError:
-        error_and_die('The maximum memory value must be greater than the minimum!')
-
-    running = is_server_running(s.server_dir)
-
-    if args.backup:
-        print_pre = '[' + Colors.yellow_green + 'live-backup' + Colors.end + '] '
-        try:
-            run_server_backup(print_pre, s, quiet)
-        except BackupFileAlreadyExistsError as e:
-            send_command('say Backup Done!', s.screen_name)
-            error_and_die(e)
-    elif args.create:
-        try:
-            create_server(s, quiet)
-        except ServerAlreadyRunningError as e:
-            error_and_die(e.msg.strip('"'))
-    elif args.genconf:
-        print(s.as_conf_file)
-    elif args.json:
-        if running:
-            s.running = running
-        print(s.as_json)
-    elif args.list:
-        if running and not quiet:
-            print_pre = '[' + Colors.yellow_green + 'list' + Colors.end + '] '
-            players = list_players(s)
-            if players:
-                for p in players:
-                    print(print_pre + Colors.light_blue + p.strip() + Colors.end)
-            else:
-                print(print_pre +
-                      Colors.yellow +
-                      'Nobody on right now :(' +
-                      Colors.end
-                      )
-        elif not quiet:
-            error_and_die("{} is not running!".format(s.world_name))
-        else:
-            die_silently()
-    elif args.offline:
-        print_pre = '[' + Colors.yellow_green + 'offline-backup' + Colors.end + '] '
-        try:
-            run_server_backup(print_pre, s, quiet, offline=True)
-        except BackupFileAlreadyExistsError as e:
-            start_server(None, s, quiet)
-            error_and_die(e)
-    elif args.say:
-        send_command('say ' + args.say, s.screen_name)
-    elif args.restart:
-        print_pre = '[' + Colors.yellow_green + 'restart' + Colors.end + '] '
-        try:
-            restart_server(print_pre, s, quiet)
-        except BackupFileAlreadyExistsError as e:
-            error_and_die(e)
-    elif args.start:
-        print_pre = '[' + Colors.yellow_green + 'start' + Colors.end + '] '
-        try:
-            start_server(print_pre, s, quiet)
-        except ServerAlreadyRunningError as e:
-            error_and_die(e)
-    elif args.stop:
-        print_pre = '[' + Colors.yellow_green + 'stop' + Colors.end + '] '
-        try:
-            stop_server(
-                print_pre,
-                s.screen_name,
-                s.server_dir,
-                s.world_name,
-                quiet
-            )
-        except ServerNotRunningError as e:
-            error_and_die(e)
-    # elif args.testing:
-    #     print(list_players_as_list(list_players(s)))
-    elif args.uptime:
-        try:
-            raw_uptime = get_uptime_raw(s.server_dir, s.world_name, quiet)
-            u = get_uptime(raw_uptime)
-            uptime_string = get_uptime_string(u)
-            print(
-                Colors.green + '{}'.format(s.world_name) +
-                Colors.blue + ' has been up for ' +
-                Colors.yellow_green + uptime_string + Colors.end
-            )
-        except ServerNotRunningError as e:
-            error_and_die(e)
-    elif args.web:
-            try:
-                run_webui(s, quiet)
-            except MinimalInstallError as e:
-                if not quiet:
-                    error_and_die(e)
-                else:
-                    die_silently()
-    else:
-        parser.print_usage()
-
-
-def main():
-    # TODO: better webui logging and logging in general
-    # TODO: get actual level seed
-    # ensure dependencies are here
-    try:
-        dependency_check(
-            get_exe_path('java'),
-            get_exe_path('screen')
-        )
-    except MissingExeError as e:
-        error_and_die(e)
-
-    # process cli args and do our stuff
-    argz = sys.argv[1:]
-    arg_parse(argz)
-
-
-if __name__ == '__main__':
-    main()
